+++ Sources/VCore/Extensions/Foundation/Collection.SafeSubscript.swift
@@ -1,5 +1,5 @@
 //
-//  Array.SafeSubscript.swift
+//  Collection.SafeSubscript.swift
 //  VCore
 //
 //  Created by Vakhtang Kontridze on 9/12/21.
@@ -7,8 +7,8 @@
 
 import Foundation
 
-// MARK: - Safe Array Subscript
-extension Array {
+// MARK: - Collection Safe Subscript
+extension Collection {
     /// Accesses the element at the specified position, but returns `nil` if out of bounds.
     ///
     ///     let numbers: [Int] = [1, 3, 5]
@@ -16,8 +16,8 @@ extension Array {
     ///     let firstNum: Int? = numbers[safe: 0] // 1
     ///     let fourthNum: Int? = numbers[safe: 4] // nil
     ///
-    public subscript(safe index: Int) -> Element? {
-        guard index >= 0, index < count else { return nil }
+    public subscript(safe index: Index) -> Element? {
+        guard indices.contains(index) else { return nil }
         return self[index]
     }
 }